refactor: move to `components/custom/job`
