+++ app/(private)/skr/tasks/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import JobList from "@/components/custom/job-list";
+import JobList from "@/components/custom/job/job-list";
 import SeekerTaskSearchBar from "@/components/custom/seeker-task-search-bar";
 
 export default function SeekerTasksPage() {
